requirements.txt


Flask~=3.0.3
face_recognition==1.3.0
numpy~=1.26.4
opencv-python~=4.10.0.82
werkzeug~=3.0.3
pillow~=10.3.0
mysql~=0.0.3
mysql-connector-python~=8.4.0